DIABLO-532, put focus on page-title when 'loading' becomes false
