+++ src/router.ts
@@ -23,9 +23,6 @@ Vue.use(Router)
 
 const router = new Router({
   mode: 'history',
-  scrollBehavior() {
-    return { x: 0, y: 0 }
-  },
   routes: [
     {
       path: '/',
